requirements.txt
pytest>=8.0.0
openai>=1.0.0
transliterate>=1.10.2
pykakasi>=2.2.1
pypinyin>=0.49.0
unidecode>=1.3.6
python-dotenv>=1.0.0
